Fix: Ask AI Remote Config keys were never ingested
ingestConfig works from an explicit allow-list, and the four ai_assistant_*
keys were not on it. AiFeatureConfig read them from AdPreferenceStore, but
nothing ever wrote them there, so every value fell through to its compiled
default. The practical effect: ai_assistant_enabled resolved to BuildConfig.DEBUG
and a RELEASE build could never be switched on, however the config was set.

Also a type mismatch that would have thrown once the key did arrive: ints are
written with putInt, and freeQueryLimit was reading with getLong.

Adds docs/remote-config-ai.md -- the four keys with types and defaults, a
paste-ready JSON block, the interaction with the four user-facing switches, and
the full 66-key list of everything the app ingests. The AI block is also added
to both audiences in docs/remote-config-GET_DATA_LIST.json.

+++ app/src/main/java/com/numify/callerid/lookup/repository/assistant/AiFeatureConfig.kt
@@ -39,7 +39,7 @@ object AiFeatureConfig {
      */
     fun freeQueryLimit(context: Context): Int =
         AdPreferenceStore.getInstance(context)
-            .getLong("${BLOCK}_free_queries", DEFAULT_FREE_QUERIES).toInt()
+            .getInt("${BLOCK}_free_queries", DEFAULT_FREE_QUERIES)
 
     /**
      * Base URL of the AI proxy. Blank until the endpoint exists, which is the
@@ -52,5 +52,5 @@ object AiFeatureConfig {
     fun showHomeTooltip(context: Context): Boolean =
         AdPreferenceStore.getInstance(context).getBoolean("${BLOCK}_home_tooltip", true)
 
-    private const val DEFAULT_FREE_QUERIES = 5L
+    private const val DEFAULT_FREE_QUERIES = 5
 }
